refactor graph_creators
Better separation of concerns between computations and plotting logic,
and switch to composition instead of inheritance for the graph creators.
Also added better type safety using Protocols, better abstract base classes
and data transfer objects.
