+++ shaketune/graph_creators/graph_creator_factory.py
@@ -4,17 +4,20 @@
 # Licensed under the GNU General Public License v3.0 (GPL-3.0)
 #
 # File: graph_creator_factory.py
-# Description: Contains a factory class to create the different graph creators.
-
+# Description: Factory class for creating graph creators
 
 from ..shaketune_config import ShakeTuneConfig
 from .graph_creator import GraphCreator
 
 
 class GraphCreatorFactory:
+    """Factory for creating graph creators"""
+
     @staticmethod
     def create_graph_creator(graph_type: str, config: ShakeTuneConfig) -> GraphCreator:
+        """Create a graph creator instance based on the type"""
+
         if creator_class := GraphCreator.registry.get(graph_type):
             return creator_class(config)
-        else:
-            raise NotImplementedError(f'Graph creator for {graph_type} not implemented!')
+
+        raise NotImplementedError(f'Graph creator for {graph_type} not implemented!')
